Simplify labRow/labCol handling
Previously, labRow/labCol were the "source of truth" for the row
and column names. This meant we had to remember to reorder them
and to assign them to dendrograms. By making the matrix itself
the authoritative source of row/col names, we don't have to worry
about maintaining these variables.